% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prac.R
\name{define_data}
\alias{define_data}
\title{Definición de datos}
\usage{
define_data()
}
\value{
Una lista con los siguientes elementos:
\itemize{
\item \code{Sample1_carbohidratos}: Vector numérico.
\item \code{Sample1_azucar}: Vector categórico.
}
}
\description{
Define los datos de las variables a analizar.
}
\details{
Esta función crea una lista con dos elementos:
\itemize{
\item \code{Sample1_carbohidratos}: Un vector numérico que contiene valores de carbohidratos.
\item \code{Sample1_azucar}: Un vector categórico que contiene categorías de azúcar (S1, S2, etc.).
}
}
