% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prac.R
\name{load_libraries}
\alias{load_libraries}
\title{Carga de librerías necesarias}
\usage{
load_libraries()
}
\value{
No devuelve ningún valor. Solo carga las librerías en el entorno de R.
}
\description{
Esta función carga las librerías necesarias para el análisis estadístico.
Las librerías incluyen funciones para cálculos de asimetría, curtosis, ajuste de distribuciones,
intervalos de confianza, pruebas no paramétricas, bootstrap y estimación bayesiana.
}
\details{
Las librerías cargadas son:
\itemize{
\item \code{e1071}: Para cálculos de asimetría y curtosis.
\item \code{MASS}: Para ajuste de distribuciones (por ejemplo, \code{fitdistr}).
\item \code{DescTools}: Para intervalos de confianza y pruebas estadísticas.
\item \code{PropCIs}: Para intervalos de confianza de proporciones.
\item \code{rcompanion}: Para pruebas no paramétricas.
\item \code{EstimationTools}: Para estimaciones puntuales y por intervalos.
\item \code{boot}: Para técnicas de bootstrap.
\item \code{BayesFactor}: Para estimación bayesiana.
}
}
