% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prac.R
\name{setup_environment}
\alias{setup_environment}
\title{Configuración inicial del entorno}
\usage{
setup_environment(seed = 2023, bootR = 1000)
}
\arguments{
\item{seed}{Valor de la semilla para reproducibilidad (por defecto: 2023).}

\item{bootR}{Número de réplicas para métodos de bootstrap (por defecto: 1000).}
}
\value{
No devuelve ningún valor. Solo configura el entorno.
}
\description{
Configura el entorno para garantizar reproducibilidad y definir parámetros comunes.
}
\details{
Esta función fija la semilla del generador de números aleatorios (\code{set.seed}) y define
el número de réplicas para métodos de bootstrap.
}
